+++ jbt/src/main/java/org/hibernate/tool/orm/jbt/internal/util/MockConnectionProvider.java
@@ -1,4 +1,4 @@
-package org.hibernate.tool.orm.jbt.util;
+package org.hibernate.tool.orm.jbt.internal.util;
 
 import java.sql.Array;
 import java.sql.Blob;
