+++ jbt/src/test/java/org/hibernate/tool/orm/jbt/api/wrp/SessionFactoryWrapperTest.java
@@ -19,7 +19,7 @@ import org.hibernate.SessionFactory;
 import org.hibernate.cfg.AvailableSettings;
 import org.hibernate.cfg.Configuration;
 import org.hibernate.tool.orm.jbt.internal.factory.SessionFactoryWrapperFactory;
-import org.hibernate.tool.orm.jbt.util.MockConnectionProvider;
+import org.hibernate.tool.orm.jbt.internal.util.MockConnectionProvider;
 import org.hibernate.tool.orm.jbt.util.MockDialect;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
