+++ jbt/src/test/java/org/hibernate/tool/orm/jbt/util/NativeConfigurationTest.java
@@ -25,6 +25,7 @@ import org.hibernate.mapping.PersistentClass;
 import org.hibernate.mapping.Table;
 import org.hibernate.tool.internal.reveng.strategy.DefaultStrategy;
 import org.hibernate.tool.orm.jbt.internal.util.MetadataHelper;
+import org.hibernate.tool.orm.jbt.internal.util.MockConnectionProvider;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.w3c.dom.Document;
